requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.15.0
requests==2.25.1
SQLAlchemy==1.3.23
urllib3==1.26.3
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
